Added account parent route and fixed links
